renamed state.js to reactstate.js
